Refactor more groupby and resample tests
xref #5409
