+++ xarray/tests/test_dataset.py
@@ -3614,19 +3614,6 @@ class TestDataset:
         expected = Dataset({"y": ("x", [0, 1, 4])}, {"z": 2, "x": [0, 1, 2]})
         assert_identical(actual, expected)
 
-        ds = Dataset({"a": ("x", range(3))}, {"b": ("x", ["A"] * 2 + ["B"])})
-        actual = ds.groupby("b").assign(c=lambda ds: 2 * ds.a)
-        expected = ds.merge({"c": ("x", [0, 2, 4])})
-        assert_identical(actual, expected)
-
-        actual = ds.groupby("b").assign(c=lambda ds: ds.a.sum())
-        expected = ds.merge({"c": ("x", [1, 1, 2])})
-        assert_identical(actual, expected)
-
-        actual = ds.groupby("b").assign_coords(c=lambda ds: ds.a.sum())
-        expected = expected.set_coords("c")
-        assert_identical(actual, expected)
-
     def test_assign_coords(self):
         ds = Dataset()
 
@@ -3757,200 +3744,6 @@ class TestDataset:
         data = Dataset({"foo": (("x",), [])}, {"x": []})
         selected = data.squeeze(drop=True)
         assert_identical(data, selected)
-
-    def test_resample_and_first(self):
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-
-        actual = ds.resample(time="1D").first(keep_attrs=True)
-        expected = ds.isel(time=[0, 4, 8])
-        assert_identical(expected, actual)
-
-        # upsampling
-        expected_time = pd.date_range("2000-01-01", freq="3H", periods=19)
-        expected = ds.reindex(time=expected_time)
-        actual = ds.resample(time="3H")
-        for how in ["mean", "sum", "first", "last"]:
-            method = getattr(actual, how)
-            result = method()
-            assert_equal(expected, result)
-        for method in [np.mean]:
-            result = actual.reduce(method)
-            assert_equal(expected, result)
-
-    def test_resample_min_count(self):
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-        # inject nan
-        ds["foo"] = xr.where(ds["foo"] > 2.0, np.nan, ds["foo"])
-
-        actual = ds.resample(time="1D").sum(min_count=1)
-        expected = xr.concat(
-            [
-                ds.isel(time=slice(i * 4, (i + 1) * 4)).sum("time", min_count=1)
-                for i in range(3)
-            ],
-            dim=actual["time"],
-        )
-        assert_equal(expected, actual)
-
-    def test_resample_by_mean_with_keep_attrs(self):
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-        ds.attrs["dsmeta"] = "dsdata"
-
-        resampled_ds = ds.resample(time="1D").mean(keep_attrs=True)
-        actual = resampled_ds["bar"].attrs
-        expected = ds["bar"].attrs
-        assert expected == actual
-
-        actual = resampled_ds.attrs
-        expected = ds.attrs
-        assert expected == actual
-
-        with pytest.warns(
-            UserWarning, match="Passing ``keep_attrs`` to ``resample`` has no effect."
-        ):
-            ds.resample(time="1D", keep_attrs=True)
-
-    def test_resample_loffset(self):
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-        ds.attrs["dsmeta"] = "dsdata"
-
-        # Our use of `loffset` may change if we align our API with pandas' changes.
-        # ref https://github.com/pydata/xarray/pull/4537
-        actual = ds.resample(time="24H", loffset="-12H").mean().bar
-        expected_ = ds.bar.to_series().resample("24H").mean()
-        expected_.index += to_offset("-12H")
-        expected = DataArray.from_series(expected_)
-        assert_allclose(actual, expected)
-
-    def test_resample_by_mean_discarding_attrs(self):
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-        ds.attrs["dsmeta"] = "dsdata"
-
-        resampled_ds = ds.resample(time="1D").mean(keep_attrs=False)
-
-        assert resampled_ds["bar"].attrs == {}
-        assert resampled_ds.attrs == {}
-
-    def test_resample_by_last_discarding_attrs(self):
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-        ds.attrs["dsmeta"] = "dsdata"
-
-        resampled_ds = ds.resample(time="1D").last(keep_attrs=False)
-
-        assert resampled_ds["bar"].attrs == {}
-        assert resampled_ds.attrs == {}
-
-    @requires_scipy
-    def test_resample_drop_nondim_coords(self):
-        xs = np.arange(6)
-        ys = np.arange(3)
-        times = pd.date_range("2000-01-01", freq="6H", periods=5)
-        data = np.tile(np.arange(5), (6, 3, 1))
-        xx, yy = np.meshgrid(xs * 5, ys * 2.5)
-        tt = np.arange(len(times), dtype=int)
-        array = DataArray(data, {"time": times, "x": xs, "y": ys}, ("x", "y", "time"))
-        xcoord = DataArray(xx.T, {"x": xs, "y": ys}, ("x", "y"))
-        ycoord = DataArray(yy.T, {"x": xs, "y": ys}, ("x", "y"))
-        tcoord = DataArray(tt, {"time": times}, ("time",))
-        ds = Dataset({"data": array, "xc": xcoord, "yc": ycoord, "tc": tcoord})
-        ds = ds.set_coords(["xc", "yc", "tc"])
-
-        # Re-sample
-        actual = ds.resample(time="12H").mean("time")
-        assert "tc" not in actual.coords
-
-        # Up-sample - filling
-        actual = ds.resample(time="1H").ffill()
-        assert "tc" not in actual.coords
-
-        # Up-sample - interpolation
-        actual = ds.resample(time="1H").interpolate("linear")
-        assert "tc" not in actual.coords
-
-    def test_resample_old_api(self):
-
-        times = pd.date_range("2000-01-01", freq="6H", periods=10)
-        ds = Dataset(
-            {
-                "foo": (["time", "x", "y"], np.random.randn(10, 5, 3)),
-                "bar": ("time", np.random.randn(10), {"meta": "data"}),
-                "time": times,
-            }
-        )
-
-        with pytest.raises(TypeError, match=r"resample\(\) no longer supports"):
-            ds.resample("1D", "time")
-
-        with pytest.raises(TypeError, match=r"resample\(\) no longer supports"):
-            ds.resample("1D", dim="time", how="mean")
-
-        with pytest.raises(TypeError, match=r"resample\(\) no longer supports"):
-            ds.resample("1D", dim="time")
-
-    def test_resample_ds_da_are_the_same(self):
-        time = pd.date_range("2000-01-01", freq="6H", periods=365 * 4)
-        ds = xr.Dataset(
-            {
-                "foo": (("time", "x"), np.random.randn(365 * 4, 5)),
-                "time": time,
-                "x": np.arange(5),
-            }
-        )
-        assert_identical(
-            ds.resample(time="M").mean()["foo"], ds.foo.resample(time="M").mean()
-        )
-
-    def test_ds_resample_apply_func_args(self):
-        def func(arg1, arg2, arg3=0.0):
-            return arg1.mean("time") + arg2 + arg3
-
-        times = pd.date_range("2000", freq="D", periods=3)
-        ds = xr.Dataset({"foo": ("time", [1.0, 1.0, 1.0]), "time": times})
-        expected = xr.Dataset({"foo": ("time", [3.0, 3.0, 3.0]), "time": times})
-        actual = ds.resample(time="D").map(func, args=(1.0,), arg3=1.0)
-        assert_identical(expected, actual)
 
     def test_to_array(self):
         ds = Dataset(
@@ -4451,24 +4244,6 @@ class TestDataset:
         expected = ds.assign_coords(c=42)
         assert_identical(expected, result)
 
-        # groupby
-        expected = Dataset({"a": ("x", range(4))}, {"x": [0, 1, 2, 3]})
-        for target in [ds, expected]:
-            target.coords["b"] = ("x", [0, 0, 1, 1])
-        actual = ds.groupby("b").fillna(DataArray([0, 2], dims="b"))
-        assert_identical(expected, actual)
-
-        actual = ds.groupby("b").fillna(Dataset({"a": ("b", [0, 2])}))
-        assert_identical(expected, actual)
-
-        # attrs with groupby
-        ds.attrs["attr"] = "ds"
-        ds.a.attrs["attr"] = "da"
-        actual = ds.groupby("b").fillna(Dataset({"a": ("b", [0, 2])}))
-        assert actual.attrs == ds.attrs
-        assert actual.a.name == "a"
-        assert actual.a.attrs == ds.a.attrs
-
         da = DataArray(range(5), name="a", attrs={"attr": "da"})
         actual = da.fillna(1)
         assert actual.name == "a"
@@ -4527,22 +4302,6 @@ class TestDataset:
         expected = Dataset({"a": (("x", "y"), [[np.nan, 1], [2, 3]])})
         actual = ds.where(ds > 0)
         assert_identical(expected, actual)
-
-        # groupby
-        ds = Dataset({"a": ("x", range(5))}, {"c": ("x", [0, 0, 1, 1, 1])})
-        cond = Dataset({"a": ("c", [True, False])})
-        expected = ds.copy(deep=True)
-        expected["a"].values = [0, 1] + [np.nan] * 3
-        actual = ds.groupby("c").where(cond)
-        assert_identical(expected, actual)
-
-        # attrs with groupby
-        ds.attrs["attr"] = "ds"
-        ds.a.attrs["attr"] = "da"
-        actual = ds.groupby("c").where(cond)
-        assert actual.attrs == ds.attrs
-        assert actual.a.name == "a"
-        assert actual.a.attrs == ds.a.attrs
 
         # attrs
         da = DataArray(range(5), name="a", attrs={"attr": "da"})
